refactor: canceled spec
