+++ tests/internal/workflow/cancelling-a-workflow-with-two-part-tariff-bill-run.spec.js
@@ -1,4 +1,4 @@
-import scenarioData from '../../support/scenarios/licence-pre-sroc-with-workflow-and-bill-run.scenario.js'
+import scenarioData from '../../support/scenarios/licence-with-workflow-and-two-part-tariff-bill-run.scenario.js'
 import { formatLongDate } from '../../support/helpers/date.helpers.js'
 import { test, expect } from '../../support/fixtures.js'
 
@@ -6,8 +6,9 @@ test.describe('Cancelling a workflow item (internal)', { tag: ['@presroc', '@sup
   let company
   let licence
 
-  test.beforeAll(async ({ setup }) => {
-    const scenario = scenarioData()
+  test.beforeAll(async ({ calculatedDates, setup }) => {
+    const dates = await calculatedDates()
+    const scenario = scenarioData(dates)
 
     const {
       licences: [scenarioLicence],
@@ -24,7 +25,7 @@ test.describe('Cancelling a workflow item (internal)', { tag: ['@presroc', '@sup
     await login(users.billingAndData)
   })
 
-  test('does not flag the licence for supplementary billing', async ({ page, users }) => {
+  test('flags the licence for supplementary billing', async ({ page, users }) => {
     await page.goto(`/system/licences/${licence.id}/summary`)
 
     await expect(page.locator('h1')).toContainText(`Licence summary ${licence.licenceRef}`)
@@ -54,6 +55,8 @@ test.describe('Cancelling a workflow item (internal)', { tag: ['@presroc', '@sup
     await expect(page.locator('h1')).toContainText('Licence set up')
     await page.getByRole('link', { name: 'Licence summary' }).click()
 
-    await expect(page.locator('.govuk-notification-banner')).not.toBeVisible()
+    await expect(page.locator('.govuk-notification-banner__content')).toContainText(
+      'This licence has been marked for the next two-part tariff supplementary bill run.'
+    )
   })
 })
