+++ tests/internal/workflow/deleting-a-licence-from-workflow.spec.js
@@ -59,3 +59,5 @@ test.describe('Deleting a licence from workflow (internal)', { tag: '@supplement
     )
   })
 })
+
+// Chnage some statsu to setup or not
